Refactor project

+++ src/main/java/com/earntogether/qlysotietkiem/model/ReportModel.java
@@ -3,5 +3,5 @@ package com.earntogether.qlysotietkiem.model;
 import lombok.Builder;
 
 @Builder
-public record ReportModel(int day, long numOfOpen, long numOfClosed) {
+public record ReportModel(int day, int numOfOpen, int numOfClosed) {
 }
